+++ src/commands/index.ts
@@ -1,3 +1,4 @@
 export * from "./setBadgeCommand";
 export * from "./setColorCommand";
 export * from "./setEmojiBadgeCommand";
+export * from "./setFilesColorCommand";