+++ src/commands/setColorFileExtCommand.ts
@@ -1,17 +1,19 @@
 import * as vscode from "vscode";
-import { getColorOptions } from "../utils";
+import { getColorOptions, getFileExtension } from "../utils";
 import { patchConfig } from "../utils/patchConfig";
 
 export const setColorFileExtCommand = (context: vscode.ExtensionContext) => {
-  return (uri: vscode.Uri, uriArray: vscode.Uri[]) => {
-    const fileExtension = uri.fsPath.split(".").pop();
+  return (uri: vscode.Uri, uriArray: vscode.Uri[] = []) => {
+    const fileExtension = getFileExtension(uri.fsPath);
 
     if (!fileExtension) {
       vscode.window.showWarningMessage("Only files can be selected");
       return;
     }
 
-    if (uriArray.length > 1) {
+    const selectedUris = uriArray.length ? uriArray : [uri];
+
+    if (selectedUris.length > 1) {
       vscode.window.showWarningMessage("Only one file can be selected");
       return;
     }